api/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
openai
supabase
pydantic
pydantic-settings
pytest
httpx
passlib[bcrypt]
email-validator
fastapi-mail
